# feat(benchmark): add a perf benchmark macro