<!--  Thanks for sending a pull request! -->
<!--  Have I run `make ci`? -->

**What type of PR is this?**
feat

**What this PR does / why we need it**:
Add a perf benchmark macro to do benchmark when assets exist.

```log
10 assets bench::perf_execute  ... bench: 109,202,563 ns/iter (+/- 6,378,009)
100 assets bench::perf_execute  ... bench: 108,859,512 ns/iter (+/- 2,977,622)
1000 assets bench::perf_execute ... bench: 108,037,404 ns/iter (+/- 4,539,634)
10000 asset bench::perf_execute  ... bench: 100,244,123 ns/iter (+/- 18,935,087)
```

It seems that performance has nothing to do with the number of existing assets.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes #

**Special notes for your reviewer**: